Removed the decode method that was using lodepng. No image decoding happens in the engine.

+++ UntoldEngine4D/Constants/CommonProtocols.h
@@ -44,12 +44,12 @@ namespace U4DEngine {
         /**
          @brief Sprite width
          */
-        int width;
+        float width;
         
         /**
          @brief Sprite height
          */
-        int height;
+        float height;
         
     }SPRITEDATA;
